refactor(devtools): move RecordedStep to dev-tools.types
RecordedStep is the one type in tutorial-exporter.ts with live consumers
(the block editor recording hooks and recorded-steps-processor). Move it to
dev-tools.types.ts alongside the other shared dev-tools types so the next
commit can delete the exporter without stranding it.

Pure type move: no behaviour change, no value-level edits. Only four sites
imported from tutorial-exporter directly; every other consumer already went
through the utils/devtools barrel and needs no change.
