+++ src/utils/devtools/action-recorder.util.ts
@@ -5,7 +5,7 @@
  * providing common operations like extracting selectors, filtering by action type, etc.
  */
 
-import type { RecordedStep } from './tutorial-exporter';
+import type { RecordedStep } from './dev-tools.types';
 import type { ExtractedSelector } from './dev-tools.types';
 
 /**
